Fixes to circular imports Moving some classes around to allow correct importing of files
