+++ escape.py
@@ -1,11 +1,14 @@
 """
 Main escape room module, executes the game and creates the base classes
 """
-import escaperoom.engine
-import escaperoom.transcript
+from escaperoom.engine import Engine
+from escaperoom.transcript import Transcript
+from escaperoom.utils import Inventory, Utils
 
 RUN_GAME = True
-transcript = escaperoom.transcript.Transcript()
+transcript = Transcript()
+utils = Utils(transcript)
+inventory = Inventory(transcript)
 transcript.print_message("Welcome to the escape room game")
 transcript.print_message(r"""
 ___________                                    __________                                                             
@@ -19,7 +22,7 @@ transcript.print_message("Enjoy the escape room!")
 transcript.print_message("Type quit to quit")
 transcript.print_message("Type hint for assistance")
 
-engine = escaperoom.engine.Engine(transcript)
+engine = Engine(transcript, inventory, utils)
 
 while RUN_GAME:
     next_step = input("What would you like to do?")
