+++ escaperoom/location.py
@@ -1,9 +1,15 @@
-"""
-currentroom.py stores the CurrentRoom enum, and methods on what items or use objects are available for a room
-"""
 from enum import Enum
 
-import escaperoom.inventory
+
+class Item(Enum):
+    """
+    The enum of items that are possible to get during the game, the default is ITEM_NOTHING
+    """
+    ITEM_DNS = "dns.cfg"
+    ITEM_VAULT = "vault_dump.txt"
+    ITEM_MALWARE = "pro_tree.jsonl"
+    ITEM_SOC = "auth.log"
+    ITEM_NOTHING = "no item"
 
 
 class CurrentRoom(Enum):
@@ -40,13 +46,12 @@ class CurrentRoom(Enum):
         return None
 
     @staticmethod
-    def get_room_item(current_room: Enum) -> escaperoom.inventory.Item:
+    def get_room_item(current_room: Enum) -> Item:
         """
         Will tell you which item is inspectable in the current room
         :param current_room: The room you want to know the item of
         :return: The item that can be inspected, or no item if none is found
         """
-        from escaperoom.inventory import Item
         match current_room:
             case CurrentRoom.DNS:
                 return Item.ITEM_DNS
